+++ modules/contracts/adapters/dex/src/exchanges/exchange_resolver.rs
@@ -24,8 +24,8 @@ pub(crate) fn identify_exchange(value: &str) -> Result<Box<dyn Identify>, DexErr
         abstract_astroport_adapter::ASTROPORT => {
             Ok(Box::<abstract_astroport_adapter::dex::Astroport>::default())
         }
-        abstract_kujira_adapter::KUJIRA => {
-            Ok(Box::<abstract_kujira_adapter::dex::Kujira>::default())
+        abstract_kujira_adapter::dex::FIN => {
+            Ok(Box::<abstract_kujira_adapter::dex::Fin>::default())
         }
         abstract_astrovault_adapter::ASTROVAULT => {
             Ok(Box::<abstract_astrovault_adapter::dex::Astrovault>::default())
@@ -56,9 +56,9 @@ pub(crate) fn resolve_exchange(value: &str) -> Result<Box<dyn DexCommand>, DexEr
         abstract_astroport_adapter::ASTROPORT => {
             Ok(Box::<abstract_astroport_adapter::dex::Astroport>::default())
         }
-        #[cfg(feature = "bow")]
-        abstract_kujira_adapter::KUJIRA => {
-            Ok(Box::<abstract_kujira_adapter::dex::Kujira>::default())
+        #[cfg(feature = "fin")]
+        abstract_kujira_adapter::dex::FIN => {
+            Ok(Box::<abstract_kujira_adapter::dex::Fin>::default())
         }
         #[cfg(feature = "astrovault")]
         abstract_astrovault_adapter::ASTROVAULT => {