+++ modules/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fixed feature namings for kujira adapters to use protocol name instead of the chain name
+
 ## [0.22.1] - 2024-05-08
 
 ### Added